mlops-template-gitlab/seedcode/mlops-gitlab-project-seedcode-model-build/test/requirements.txt
awscli
boto3
pyyaml
pandas